refactor(config): promote McpConnectionAuth into @executor/config
The runtime-shape auth type lived in @executor/plugin-mcp while its
file-shape counterpart (McpAuthConfig) lived in @executor/config.
Splitting the pair across packages made the plugin import
SECRET_REF_PREFIX back from @executor/config just to serialize auth —
an inversion that made the forward and inverse translators hard to
keep in sync.

Co-locate both shapes next to each other in @executor/config/schema.ts.
The plugin re-exports McpConnectionAuth from its types barrel so
existing downstream consumers (apps/local migrate-connections,
apps/cloud migrate-mcp-connections) are unchanged.

No behavior change.
